#29
Pass.js
----------
Added check in expiration function to permit date type of Date as well as string
Added check for date type into relevance function (no date check previously)

Utils.js
----------
Added a similar date type check to that in pass.js
Modified creation of parseDate to skip the parse step if date is a Date rather than a string.

Possible improvement: As it stands, [I suspect that] you cannot pass in a Date as a Moment object.
A small change to the date check should permit that.