#26
fieldsArray.js
-----------------
I've added a wrapper class around the existing one. This will now hold the filedKeys set so that is it shared across all the array instances, but within an instance of the current pass. Other passes will have their own instance, and hence their own set of keys.

pass.js
----------
A small change required to the instantiation of fieldsArray and the way the data is copied into the pass data.